+++ src/Pixel.Automation.Designer.ViewModels/Factory/VersionManagerFactory.cs
@@ -16,6 +16,7 @@ namespace Pixel.Automation.Designer.ViewModels.Factory
 {
     public class VersionManagerFactory : IVersionManagerFactory
     {
+        private readonly IDataManagerFactory projectAssetsDataManagerFactory;
         private readonly IWorkspaceManagerFactory workspaceManagerFactory;
         private readonly IReferenceManagerFactory referenceManagerFactory;
         private readonly ISerializer serializer;
@@ -24,22 +25,25 @@ namespace Pixel.Automation.Designer.ViewModels.Factory
         private readonly IProjectDataManager projectDataManager;
         private readonly INotificationManager notificationManager;
         private readonly IEventAggregator eventAggregator;
+        private readonly IApplicationFileSystem applicationFileSystem;
         private readonly ApplicationSettings applicationSettings;
 
         public VersionManagerFactory(ISerializer serializer, IWorkspaceManagerFactory workspaceManagerFactory, 
             IReferenceManagerFactory referenceManagerFactory, IApplicationDataManager applicationDataManager,
-            IPrefabDataManager prefabDataManager, IProjectDataManager projectDataManager, IEventAggregator eventAggregator,
-            INotificationManager notificationManager, ApplicationSettings applicationSettings)
+            IPrefabDataManager prefabDataManager, IProjectDataManager projectDataManager, IDataManagerFactory projectAssetsDataManagerFactory,
+            IEventAggregator eventAggregator, INotificationManager notificationManager, ApplicationSettings applicationSettings, IApplicationFileSystem applicationFileSystem)
         {
             this.serializer = Guard.Argument(serializer, nameof(serializer)).NotNull().Value;
             this.workspaceManagerFactory = Guard.Argument(workspaceManagerFactory, nameof(workspaceManagerFactory)).NotNull().Value;
             this.referenceManagerFactory = Guard.Argument(referenceManagerFactory, nameof(referenceManagerFactory)).NotNull().Value;
             this.applicationDataManager = Guard.Argument(applicationDataManager, nameof(applicationDataManager)).NotNull().Value;
             this.projectDataManager = Guard.Argument(projectDataManager, nameof(projectDataManager)).NotNull().Value;
+            this.projectAssetsDataManagerFactory = Guard.Argument(projectAssetsDataManagerFactory, nameof(projectAssetsDataManagerFactory)).NotNull().Value;
             this.prefabDataManager = Guard.Argument(prefabDataManager, nameof(prefabDataManager)).NotNull().Value;
             this.eventAggregator = Guard.Argument(eventAggregator, nameof(eventAggregator)).NotNull().Value;
             this.notificationManager = Guard.Argument(notificationManager, nameof(notificationManager)).NotNull().Value;
             this.applicationSettings = Guard.Argument(applicationSettings, nameof(applicationSettings)).NotNull();
+            this.applicationFileSystem = Guard.Argument(applicationFileSystem, nameof(applicationFileSystem)).NotNull().Value;
         }
 
         public IVersionManager CreatePrefabVersionManager(PrefabProject prefabProject)
@@ -50,7 +54,8 @@ namespace Pixel.Automation.Designer.ViewModels.Factory
 
         public IVersionManager CreateProjectVersionManager(AutomationProject automationProject)
         {           
-            return new ProjectVersionManagerViewModel(automationProject, this.workspaceManagerFactory, this.referenceManagerFactory, this.serializer, this.projectDataManager, 
+            return new ProjectVersionManagerViewModel(automationProject, this.workspaceManagerFactory, this.referenceManagerFactory, 
+                this.applicationDataManager, this.projectDataManager, this.projectAssetsDataManagerFactory, this.prefabDataManager, this.serializer, applicationFileSystem,
                this.eventAggregator, this.notificationManager, this.applicationSettings);
         }
